refactor: move functions to endpoint file
