+++ pkg/grpc/service.go
@@ -14,8 +14,8 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/cloudforet-io/cfctl/cmd/other"
 	"github.com/cloudforet-io/cfctl/pkg/format"
+	"github.com/cloudforet-io/cfctl/pkg/rest"
 	"github.com/eiannone/keyboard"
 	"github.com/spf13/viper"
 
@@ -168,13 +168,13 @@ func FetchService(serviceName string, verb string, resourceName string, options 
 	if config.Environment == "local" {
 		hostPort = strings.TrimPrefix(config.Environments[config.Environment].Endpoint, "grpc://")
 	} else {
-		apiEndpoint, err = other.GetAPIEndpoint(config.Environments[config.Environment].Endpoint)
+		apiEndpoint, err = rest.GetAPIEndpoint(config.Environments[config.Environment].Endpoint)
 		if err != nil {
 			pterm.Error.Printf("Failed to get API endpoint: %v\n", err)
 			os.Exit(1)
 		}
 		// Get identity service endpoint
-		identityEndpoint, hasIdentityService, err = other.GetIdentityEndpoint(apiEndpoint)
+		identityEndpoint, hasIdentityService, err = rest.GetIdentityEndpoint(apiEndpoint)
 		if err != nil {
 			pterm.Error.Printf("Failed to get identity endpoint: %v\n", err)
 			os.Exit(1)
